# StoreGateway: Optimize filtering users during intial sync