@harry671003 2023-11-24
This is no longer required if the `tenant-delete-mark.json` can be moved outside the tenant dir.
https://github.com/cortexproject/cortex/pull/5676